refactor(user-service): reorganizes the structure of the packages
