+++ src/main/java/com/manager/freelancer_management_api/domain/global/exceptions/BusinessException.java
@@ -1,4 +1,4 @@
-package com.manager.freelancer_management_api.domain.exceptions;
+package com.manager.freelancer_management_api.domain.global.exceptions;
 
 public class BusinessException extends RuntimeException{
     public BusinessException(String message) {
