+++ src/main/java/com/manager/freelancer_management_api/domain/global/exceptions/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package com.manager.freelancer_management_api.domain.exceptions;
+package com.manager.freelancer_management_api.domain.global.exceptions;
 
 public class InvalidTokenException extends RuntimeException {
     public InvalidTokenException(){
