+++ src/main/java/com/manager/freelancer_management_api/domain/global/exceptions/PublicKeyLoadException.java
@@ -1,4 +1,4 @@
-package com.manager.freelancer_management_api.domain.exceptions;
+package com.manager.freelancer_management_api.domain.global.exceptions;
 
 public class PublicKeyLoadException extends RuntimeException {
     public PublicKeyLoadException(String message) {
